[DSCP-105] Move HasLens to base
