+++ code/network/lib/Loot/Network/ZMQ/Instance.hs
@@ -6,8 +6,8 @@
 
 module Loot.Network.ZMQ.Instance () where
 
+import Loot.Base.HasLens (HasLens')
 import qualified Loot.Network.Class as C
-import Loot.Network.Utils (HasLens')
 import qualified Loot.Network.ZMQ.Client as ZC
 import Loot.Network.ZMQ.Common (ZTGlobalEnv, ZTNodeId, ZmqTcp)
 import qualified Loot.Network.ZMQ.Server as ZS
