+++ code/network/package.yaml
@@ -14,6 +14,7 @@ library:
     - dependent-sum
     - lens
     - lifted-async
+    - loot-base
     - loot-log
     - mtl
     - network
